requirements.txt
bokeh==2.2.2
certifi==2020.6.20
chardet==3.0.4
Flask==1.1.2
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.2
packaging==20.4
pandas==1.1.3
Pillow==8.0.0
pip==19.0.3
psycopg2-binary==2.8.5
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
setuptools==40.8.0
six==1.15.0
SQLAlchemy==1.3.20
tornado==6.0.4
typing-extensions==3.7.4.3
urllib3==1.25.10
git+https://github.com/materialscloud/mz-bokeh-package.git@v0.0.1
click==7.1.2
crossrefapi==1.5.0
itsdangerous==1.1.0
Werkzeug==1.0.1
xlrd==1.2.0
et-xmlfile==1.0.1
jdcal==1.4.1
openpyxl==3.0.1
python-dotenv==0.15.0
